move LoginLearnMore and SsoLearnMore actions to LearnMoreAction
