+++ jetbrains-core/src/software/aws/toolkits/jetbrains/services/codewhisperer/actions/CodeWhispererLearnMoreAction.kt
@@ -4,11 +4,13 @@
 package software.aws.toolkits.jetbrains.services.codewhisperer.actions
 
 import com.intellij.icons.AllIcons
+import com.intellij.ide.BrowserUtil
 import com.intellij.openapi.actionSystem.AnAction
 import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.project.DumbAware
 import software.aws.toolkits.jetbrains.services.codewhisperer.explorer.CodeWhispererExplorerActionManager
 import software.aws.toolkits.resources.message
+import java.net.URI
 
 class CodeWhispererLearnMoreAction :
     AnAction(
@@ -20,5 +22,17 @@ class CodeWhispererLearnMoreAction :
 
     override fun actionPerformed(e: AnActionEvent) {
         CodeWhispererExplorerActionManager.getInstance().showWhatIsCodeWhisperer()
+    }
+}
+
+class CodeWhispererSsoLearnMoreAction : AnAction(message("aws.settings.learn_more")), DumbAware {
+    override fun actionPerformed(e: AnActionEvent) {
+        BrowserUtil.browse(URI("https://docs.aws.amazon.com/toolkit-for-jetbrains/latest/userguide/codewhisperer.html"))
+    }
+}
+
+class CodeWhispererLoginLearnMoreAction : AnAction(message("aws.settings.learn_more")), DumbAware {
+    override fun actionPerformed(e: AnActionEvent) {
+        BrowserUtil.browse(URI("https://docs.aws.amazon.com/toolkit-for-jetbrains/latest/userguide/codewhisper-setup-general.html"))
     }
 }
